batching does not improve save speed

+++ ppa/archive/management/commands/generate_textcorpus.py
@@ -12,7 +12,8 @@ from parasolr.django import SolrQuerySet
 from progressbar import progressbar
 import logging
 import gzip
-
+import orjsonl
+from functools import cached_property
 
 class Command(BaseCommand):
     """Custom manage command to generate a text corpus from text indexed in Solr"""
@@ -42,37 +43,46 @@ class Command(BaseCommand):
             "The default of -1 considers ALL documents.",
         )
         parser.add_argument(
-            "--batch",
+            "--batch-size",
             type=int,
-            default=100,
+            default=1000,
             help="Number of docs to save at one time",
         )
+        parser.add_argument(
+            "--dry-run",
+            action='store_true',
+            help="Do not save anything, just iterate over solr",
+        )
+
+    @cached_property
+    def query_set(self):
+        return SolrQuerySet()
 
     def iter_solr(self, batch_size=1000, item_type='page', lim=None, progress=True):
         """
         Iterate over solr documents of a certain `item_type`
         """
-        i=0
-        qset = SolrQuerySet()
-        
 
-        def get_query(order=True):
-            q=qset.search(item_type=item_type)
+        def query(order=True):
+            q=self.query_set.search(item_type=item_type)
             if order: q=q.order_by('id')
             if item_type=='page': q=q.only(*self.PAGE_FIELDLIST)
             return q
 
-        q=get_query(order=False)
+        q=query(order=True)
         total = q.count()
         if lim and int(total)>lim: total=lim
-        iterr = range(0, total, batch_size)
-        if progress:
-            iterr = progressbar(iterr)
-        for step in iterr:
-            q=get_query(order=True)
-            q.set_limits(step, step+batch_size)
-            for d in q:
-                yield d
+        if batch_size>total: batch_size=total
+
+        def iter_func(): 
+            steps = range(0, total, batch_size)
+            for step in steps:
+                yield from q[step:step+batch_size]
+                # if '_result_cache' in q.__dict__: del q.__dict__['_result_cache']
+
+        iterr = progressbar(iter_func(), max_value=total) if progress else iter_func()
+        yield from iterr
+
 
     def handle(self, *args, **options):
         """
@@ -80,13 +90,15 @@ class Command(BaseCommand):
         """
         # options
         path = options['path']
+        is_dry_run = options['dry_run']
         doclimit = options['doc_limit'] if options['doc_limit']>0 else None
         progress = options['verbosity']>0
-        batch_size = options['batch']
+        batch_size = options['batch_size']
         by_batch = batch_size > 1
 
         # paths
-        os.makedirs(path, exist_ok=True)
+        if not is_dry_run:
+            os.makedirs(path, exist_ok=True)
         path_meta = os.path.join(path,'metadata.json')
         path_texts = os.path.join(path,'pages.jsonl.gz')
         
@@ -96,13 +108,14 @@ class Command(BaseCommand):
                 item_type='work',
                 lim=doclimit,
                 progress=progress,
-                # batch_size=batch_size if by_batch else 1000
+                batch_size=batch_size if by_batch else 1000
             )
 
         output_ld = list(iter_works())
-        with open(path_meta,'w') as of:
-            json.dump(output_ld, of, indent=2)
-        
+        if not is_dry_run:
+            with open(path_meta,'w') as of:
+                json.dump(output_ld, of, indent=2)
+            
 
         # save pages
         def iter_pages():
@@ -110,13 +123,18 @@ class Command(BaseCommand):
                 item_type='page',
                 lim=doclimit,
                 progress=progress,
-                # batch_size=batch_size if by_batch else 1000
+                batch_size=batch_size if by_batch else 1000
             )
         ### save pages
-        if not by_batch:
-            with gzip.open(path_texts,'wt',encoding='utf-8') as of:
-                for d in iter_pages():
-                    of.write(json.dumps(d)+'\n')
+        if is_dry_run:
+            print('dry run')
+            for x in iter_pages():
+                pass
+        elif not by_batch:
+            orjsonl.save(path_texts, iter_pages())
+            # with gzip.open(path_texts,'wt',encoding='utf-8') as of:
+            #     for d in iter_pages():
+            #         of.write(json.dumps(d)+'\n')
         else:    
             with gzip.open(path_texts,'wt',encoding='utf-8') as of:
                 batch=[]
